MST-46: Accounting report with opening and closing balance (#42)
* MST-46: Accounting report without opening and closing balance

* MST-46: Add opening and closing balance

* MST-46: Rename accounting function

* MST-46: Add documentation

* MST-46: Remove orderBy statement

* MST-46: Fix status parameter

* MST-46: Fix tests

* MST-46: Add more checks

* MST-46: yet another test fix

* MST-46: Blyat!

* MST-46: Remove retarded java timestamp

* MST-46: Bump mst version

+++ src/main/resources/db/migration/V1__init.sql
@@ -20,7 +20,7 @@ create table mst.payment (
   event_id bigint not null,
   invoice_id character varying not null,
   merchant_id character varying not null,
-  shop_id bigint not null,
+  shop_id int not null,
   customer_id character varying,
   masked_pan character varying not null,
   status character varying not null,
